Fix TOC header (#168)
In English, the translation that is supposed to be shown on top of the table of contents looks like this:

```toml
[whatsInThis]
other = "What's in this {{.Type }}"
```

When it is invoked, there are two mistakes: `{{ i18n "whatsInThis" humanize .Type }}`.

The first mistake is that you cannot call `humanize` in this context. The more severe one is that if you remove `humanize` it will pass the `.Type` directly to the i18n and that tries to extract `.Type` from it.