Use environment introduced in Hugo 0.53
The Ananke theme got environment support before Hugo did and when Hugo
added it in version 0.53 it was in a slightly different way. Running
"hugo server" will set the environment to "development" and running just
"hugo" will set it to "production". It can be overridden by OS
environment variable "HUGO_ENVIRONMENT" (note, not "HUGO_ENV") or by
using the --environment option.

Use "hugo.Environment" to detect environment instead and update the
README.