Remove scrollbars unless needed
Ananke has a very minimal design but code sections were including
horizontal and vertical scrollbars even when there is only a single line
of code. I feel that removing them unless necessary leads to a less
cluttered look with no downsides.

This also includes a rebuild of the production file as requested here:
https://github.com/budparr/gohugo-theme-ananke/pull/279#issuecomment-638253759