Allow override of date format
Previously the default date format (`January 2, 2006` ) was used for
blog and single page date visualizations.

The new parameter `date_format` can be used to override this format.
This allows the visualization of localized dates or a changed precision
(down to the minute or up to the month or year of publication).