theme(fix): add show_recent_posts parameter to stop showing recent posts on index (#795)
<!---

Add the issue number that is discussed and fixed by this PR (In the form
`Closes #123`). If this PR doesn't fix an issue, remove the line below.
This will
also lead to us not treating this PR as an important one. It might be
closed
without a review.

If there is no issue associated with this PR and you are not a
maintainer of
this repository, your PR might be closed without a review.

-->

Closes #794

You can no longer set `recent_posts_number = 0` because in current hugo
versions `{{ $n_posts := $.Param "recent_posts_number" | default 3 }}`
will default to three if recent_posts_number is 0. This used to work a
couple of versions back, I am assuming this is not a regression but
rather wanted behavior in hugo -- so I introduce a new parameter
(boolean) `show_recent_posts` to enable(default) or disable recent posts
on index.



<!---

Explain what this PR does and what existing problem it solves. If this
PR is a
work in progress, please prefix the title with [WIP].

-->

<!--

Make sure that the code is readable and well-documented. If you have
added new
functionality, please add the necessary documentation.
If testing of the new functionality is possible, please add tests.

-->